+++ .eslintrc.json
@@ -18,13 +18,11 @@
   },
   "plugins": [
     "@typescript-eslint",
-    "jest",
     "prettier",
     "react"
   ],
   "extends": [
     "eslint:recommended",
-    "plugin:jest/recommended",
     "plugin:prettier/recommended",
     "prettier/@typescript-eslint"
   ],
