+++ package.json
@@ -50,31 +50,30 @@
     "prebuild": "npm run clean",
     "build": "rollup -c rollup.config.js",
     "clean": "shx rm -rf dist",
-    "debug": "node --inspect-brk ./node_modules/.bin/jest --runInBand --rootDir src",
     "lint": "eslint --ext .js,.jsx,.ts,.tsx .",
     "preshipit": "npm run typecheck && npm run test && npm run build",
     "shipit": "cd dist && npm publish",
-    "test": "jest --rootDir src",
+    "test": "playwright-test --runner uvu test",
     "typecheck": "tsc --noEmit"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.20",
+    "@types/sinon": "^10.0.12",
     "@typescript-eslint/eslint-plugin": "^4.14.0",
     "@typescript-eslint/parser": "^4.14.0",
     "eslint": "^7.18.0",
     "eslint-config-prettier": "^7.2.0",
-    "eslint-plugin-jest": "^24.1.3",
     "eslint-plugin-prettier": "^3.3.1",
     "eslint-plugin-react": "^7.22.0",
-    "jest": "^26.6.3",
     "magic-string": "^0.25.7",
+    "playwright-test": "^8.1.1",
     "prettier": "^2.2.1",
     "rollup": "^2.38.0",
     "rollup-plugin-typescript2": "^0.29.0",
     "shx": "^0.3.3",
-    "ts-jest": "^26.4.4",
+    "sinon": "^14.0.0",
     "ts-node": "^9.1.1",
-    "typescript": "^4.1.2"
+    "typescript": "^4.1.2",
+    "uvu": "^0.5.6"
   },
   "publishConfig": {
     "access": "public"
