+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["src"],
+  "include": ["src", "test"],
   "compilerOptions": {
     "target": "ESNEXT",
     "module": "ESNEXT",
